+++ src/Deedle.Interactive/AssemblyInfo.fs
@@ -2,18 +2,18 @@
 namespace System
 open System.Reflection
 
-[<assembly: AssemblyTitleAttribute("Deedle.Math")>]
-[<assembly: AssemblyProductAttribute("Deedle.Math")>]
-[<assembly: AssemblyDescriptionAttribute("Deedle interop with Math.Net")>]
+[<assembly: AssemblyTitleAttribute("Deedle.RPlugin")>]
+[<assembly: AssemblyProductAttribute("Deedle.RPlugin")>]
+[<assembly: AssemblyDescriptionAttribute("Easy to use .NET library for data manipulation with R project integration")>]
 [<assembly: AssemblyVersionAttribute("3.0.0")>]
 [<assembly: AssemblyInformationalVersionAttribute("3.0.0-beta.1")>]
 [<assembly: AssemblyFileVersionAttribute("3.0.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] AssemblyTitle = "Deedle.Math"
-    let [<Literal>] AssemblyProduct = "Deedle.Math"
-    let [<Literal>] AssemblyDescription = "Deedle interop with Math.Net"
+    let [<Literal>] AssemblyTitle = "Deedle.RPlugin"
+    let [<Literal>] AssemblyProduct = "Deedle.RPlugin"
+    let [<Literal>] AssemblyDescription = "Easy to use .NET library for data manipulation with R project integration"
     let [<Literal>] AssemblyVersion = "3.0.0"
     let [<Literal>] AssemblyInformationalVersion = "3.0.0-beta.1"
     let [<Literal>] AssemblyFileVersion = "3.0.0"