+++ src/Deedle/Series.fs
@@ -1,4 +1,4 @@
-﻿namespace Deedle
+namespace Deedle
 
 open System
 open System.ComponentModel
@@ -1133,7 +1133,10 @@ and
   /// Shows the series content in a human-readable format. The resulting string
   /// shows a limited number of values from the series.
   member series.Format() =
-    series.Format(Formatting.StartItemCount, Formatting.EndItemCount)
+    series.Format(Formatting.RowStartItemCount, Formatting.RowEndItemCount, false)
+
+  member series.Format(showInfo) =
+    series.Format(Formatting.RowStartItemCount, Formatting.RowEndItemCount, showInfo)
 
   /// Shows the series content in a human-readable format. The resulting string
   /// shows a limited number of values from the series.
@@ -1143,7 +1146,7 @@ and
   ///    at most `itemCount/2` items at the beginning and ending of the series.
   member series.Format(itemCount) =
     let half = itemCount / 2
-    series.Format(half, half)
+    series.Format(half, half, false)
 
   /// Shows the series content in a human-readable format. The resulting string
   /// shows a limited number of values from the series.
@@ -1151,7 +1154,7 @@ and
   /// ## Parameters
   ///  - `startCount` - The number of elements to show at the beginning of the series
   ///  - `endCount` - The number of elements to show at the end of the series
-  member series.Format(startCount, endCount) =
+  member series.FormatStrings(startCount, endCount) : string [] [] =
     let getLevel ordered previous reset maxLevel level (key:'K) =
       let levelKey =
         if level = 0 && maxLevel = 0 then box key
@@ -1159,9 +1162,9 @@ and
       if ordered && (Some levelKey = !previous) then ""
       else previous := Some levelKey; reset(); levelKey.ToString()
 
-    if vector.SuppressPrinting then "(Suppressed)" else
+    if vector.SuppressPrinting then [|[|"(Suppressed)"|]|] else
       if series.IsEmpty then
-        "(Empty)"
+        [|[|"(Empty)"|]|]
       else
         let firstKey = series.GetKeyAt(0)
         let levels = CustomKey.Get(firstKey).Levels
@@ -1169,25 +1172,52 @@ and
         let reset i () = for j in i + 1 .. levels - 1 do previous.[j] := None
 
         series.GetPrintedObservations(startCount, endCount)
-        |> Seq.map (function
+        |> Array.ofSeq
+        |> Array.map (function
             | Choice1Of3(k, v) | Choice3Of3(k, v) ->
-                [ // Yield all row keys
+                [| // Yield all row keys
                   for level in 0 .. levels - 1 do
                     yield getLevel series.Index.IsOrdered previous.[level] (reset level) levels level k
                   yield "->"
-                  yield v.ToString() ]
+                  yield v.ToString()
+                |]
             | Choice2Of3() ->
-                [ yield "..."
+                [|
+                  yield "..."
                   for level in 1 .. levels - 1 do yield ""
                   yield "->"
-                  yield "..." ] )
+                  yield "..."
+                |] )
+
+
+  member series.Format(startCount, endCount, showInfo) =
+    try
+      let formattedtable =
+        series.FormatStrings(startCount, endCount)
         |> array2D
         |> Formatting.formatTable
+      if showInfo then
+        let seriesInfo = sprintf "series of %i items with %i missing values" series.KeyCount (series.KeyCount - series.ValueCount)
+        sprintf "%s%s%s" formattedtable System.Environment.NewLine seriesInfo
+      else
+        formattedtable
+    with e -> sprintf "Formatting failed: %A" e
+
 
   interface IFsiFormattable with
     member x.Format() = (x :> Series<_, _>).Format()
+    member x.FormatWithInfo() = (x :> Series<_, _>).Format(true)
 
-
+  interface ISeriesFormattable with
+    member x.InteractiveFormat(count) = (x :> Series<_, _>).FormatStrings((count/2),(count/2))
+    member x.GetValueCount() = x.ValueCount
+    member x.GetKeyCount() = x.KeyCount
+    member x.GetKeyLevels() =
+      if x.IsEmpty then
+        1
+      else
+        let firstKey = x.GetKeyAt(0)
+        CustomKey.Get(firstKey).Levels
 
   // ----------------------------------------------------------------------------------------------
   // Nicer constructor